requirements.txt
streamlit
python_docx
PyMuPDF
sentence-transformers 
pandas
nltk